Move persisting logic to CampaignDb
